+++ packages/web3-types/src/eth_types.ts
@@ -226,7 +226,7 @@ export interface Filter {
 	readonly fromBlock?: BlockNumberOrTag;
 	readonly toBlock?: BlockNumberOrTag;
 	readonly address?: Address | Address[];
-
+	readonly blockHash?: Address;
 	// Using "null" type intentionally to match specifications
 	// eslint-disable-next-line @typescript-eslint/ban-types
 	readonly topics?: (null | Topic | Topic[])[];