+++ packages/web3-types/CHANGELOG.md
@@ -135,3 +135,7 @@ Documentation:
 -   Fixed bug #6185, now web3.js compiles on typescript v5 (#6195)
 
 ## [Unreleased]
+
+### Fixed
+
+-   type `Filter` includes `blockHash` (#6206)